style: SpotBugs P1 `Reliance on default encoding ( DM_DEFAULT_ENCODING )`
- Global.<static initializer for Global>()
- Benchmark.benchmarkKanziInternal(String, String[])
- Benchmark.benchmarkKanziInternal(String, String[])
- TextCodec.<static initializer for TextCodec>()
- LyndonWords.getPositions(String, Charset)
- LyndonWords.split(String, Charset)

Found a call to a method which will perform a byte to String (or String to byte) conversion, and will assume that the default platform encoding is suitable. This will cause the application behavior to vary between platforms. Use an alternative API and specify a charset name or Charset object explicitly.

See CWE-173: Improper Handling of Alternate Encoding.